requirements.txt
streamlit
pdf2image
google-generativeai
Pillow